// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      manhwas: {
        Row: {
          author: string | null
          cover_url: string | null
          created_at: string
          featured: boolean | null
          genres: string[] | null
          id: string
          status: string
          synopsis: string | null
          title: string
          updated_at: string
          view_count: number
        }
        Insert: {
          author?: string | null
          cover_url?: string | null
          created_at?: string
          featured?: boolean | null
          genres?: string[] | null
          id?: string
          status?: string
          synopsis?: string | null
          title: string
          updated_at?: string
          view_count?: number
        }
        Update: {
          author?: string | null
          cover_url?: string | null
          created_at?: string
          featured?: boolean | null
          genres?: string[] | null
          id?: string
          status?: string
          synopsis?: string | null
          title?: string
          updated_at?: string
          view_count?: number
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      has_role: {
        Args: {
          _role: Database["public"]["Enums"]["app_role"]
          _user_id: string
        }
        Returns: boolean
      }
    }
    Enums: {
      app_role: "admin" | "user"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}